+++ .github/workflows/eftimie-00.yml
@@ -23,6 +23,7 @@ jobs:
       - uses: actions/setup-python@v5
         with:
           python-version: '3.12.7' 
-      - run: python --version
+      - run: python --versionù
+      - run: pip install -r eftimie/requirements.txt
       - run: python inizio.py
         working-directory: eftimie
