+++ eftimie/requirements.txt
@@ -1,0 +1,65 @@
+asttokens==3.0.0
+bleach==6.2.0
+certifi==2025.4.26
+charset-normalizer==3.4.1
+colorama==0.4.6
+comm==0.2.2
+contourpy==1.3.1
+cycler==0.12.1
+debugpy==1.8.13
+decorator==5.2.1
+executing==2.2.0
+fonttools==4.56.0
+idna==3.10
+ipykernel==6.29.5
+ipython==9.0.2
+ipython_pygments_lexers==1.1.1
+jedi==0.19.2
+joblib==1.4.2
+jupyter_client==8.6.3
+jupyter_core==5.7.2
+kaggle==1.7.4.2
+kagglehub==0.3.12
+kiwisolver==1.4.8
+matplotlib==3.10.1
+matplotlib-inline==0.1.7
+narwhals==1.34.1
+nest-asyncio==1.6.0
+numpy==2.2.4
+packaging==24.2
+pandas==2.2.3
+parso==0.8.4
+patsy==1.0.1
+pillow==11.1.0
+platformdirs==4.3.7
+plotly==6.0.1
+plotly-express==0.4.1
+prompt_toolkit==3.0.50
+protobuf==6.30.2
+psutil==7.0.0
+pure_eval==0.2.3
+Pygments==2.19.1
+pyparsing==3.2.3
+python-dateutil==2.9.0.post0
+python-slugify==8.0.4
+pytz==2025.2
+pywin32==310
+PyYAML==6.0.2
+pyzmq==26.3.0
+requests==2.32.3
+scikit-learn==1.6.1
+scipy==1.15.2
+seaborn==0.13.2
+setuptools==80.0.0
+six==1.17.0
+stack-data==0.6.3
+statsmodels==0.14.4
+text-unidecode==1.3
+threadpoolctl==3.6.0
+tornado==6.4.2
+tqdm==4.67.1
+traitlets==5.14.3
+tzdata==2025.2
+urllib3==2.4.0
+wcwidth==0.2.13
+webencodings==0.5.1
